major refactoring

+++ include/types/Color.hpp
@@ -7,17 +7,17 @@ struct Color {
     int b;
 
     void operator*=(double const other) {
-        r*=other;
-        g*=other;
-        b*=other;
+        r *= other;
+        g *= other;
+        b *= other;
     }
 
     Color operator*(double const other) const {
-        return {r*other, g*other, b*other};
+        return {static_cast<int>(r * other), static_cast<int>(g * other), static_cast<int>(b * other)};
     }
 
     Color operator+(Color const other) const {
-        return {r+other.r, g+other.g, b+other.b};
+        return {r + other.r, g + other.g, b + other.b};
     }
 };
